chore(angular): e2e improvements
