--- e2e/angular-core/src/angular-core.test.ts
@@ -1,105 +1,0 @@
-import {
-  checkFilesExist,
-  expectTestsPass,
-  getSelectedPackageManager,
-  getSize,
-  killPorts,
-  newProject,
-  cleanupProject,
-  runCLI,
-  runCLIAsync,
-  tmpProjPath,
-  uniq,
-  updateFile,
-  runCypressTests,
-  removeFile,
-  checkFilesDoNotExist,
-  isNotWindows,
-} from '@nrwl/e2e/utils';
-
-import { names } from '@nrwl/devkit';
-
-describe('Angular Package', () => {
-  describe('core', () => {
-    let proj: string;
-
-    beforeAll(() => (proj = newProject()));
-    afterAll(() => cleanupProject());
-
-    it('should generate an app, a lib, link them, build and test both correctly', async () => {
-      const myapp = uniq('myapp');
-      const myapp2 = uniq('myapp2');
-      const mylib = uniq('mylib');
-      runCLI(
-        `generate @nrwl/angular:app ${myapp} --directory=myDir --no-interactive`
-      );
-      runCLI(
-        `generate @nrwl/angular:app ${myapp2} --directory=myDir --no-interactive`
-      );
-      runCLI(
-        `generate @nrwl/angular:lib ${mylib} --directory=myDir --add-module-spec --no-interactive`
-      );
-
-      updateFile(
-        `apps/my-dir/${myapp}/src/app/app.module.ts`,
-        `
-        import { NgModule } from '@angular/core';
-        import { BrowserModule } from '@angular/platform-browser';
-        import { MyDir${
-          names(mylib).className
-        }Module } from '@${proj}/my-dir/${mylib}';
-        import { AppComponent } from './app.component';
-        import { NxWelcomeComponent } from './nx-welcome.component';
-
-        @NgModule({
-          imports: [BrowserModule, MyDir${names(mylib).className}Module],
-          declarations: [AppComponent, NxWelcomeComponent],
-          bootstrap: [AppComponent]
-        })
-        export class AppModule {}
-      `
-      );
-      runCLI(
-        `run-many --target build --projects=my-dir-${myapp},my-dir-${myapp2} --parallel --prod --output-hashing none`
-      );
-
-      checkFilesExist(`dist/apps/my-dir/${myapp}/main.js`);
-
-      // This is a loose requirement because there are a lot of
-      // influences external from this project that affect this.
-      const es2015BundleSize = getSize(
-        tmpProjPath(`dist/apps/my-dir/${myapp}/main.js`)
-      );
-      console.log(
-        `The current es2015 bundle size is ${es2015BundleSize / 1000} KB`
-      );
-      expect(es2015BundleSize).toBeLessThanOrEqual(160000);
-
-      runCLI(
-        `run-many --target test --projects=my-dir-${myapp},my-dir-${mylib} --parallel`
-      );
-
-      if (runCypressTests()) {
-        const e2eResults = runCLI(`e2e my-dir-${myapp}-e2e --no-watch`);
-        expect(e2eResults).toContain('All specs passed!');
-        expect(await killPorts()).toBeTruthy();
-      }
-    }, 1000000);
-
-    it('should support workspaces w/o workspace config file', async () => {
-      if (isNotWindows()) {
-        removeFile('workspace.json');
-        const myapp = uniq('myapp');
-        runCLI(
-          `generate @nrwl/angular:app ${myapp} --directory=myDir --routing`
-        );
-
-        runCLI(`build my-dir-${myapp} --aot`);
-        expectTestsPass(await runCLIAsync(`test my-dir-${myapp} --no-watch`));
-        expect(() =>
-          checkFilesDoNotExist('workspace.json', 'angular.json')
-        ).not.toThrow();
-      }
-    }, 1000000);
-  });
-});
